IMS: Fix manage conference participants missing issue after SRVCC.
Contact cache entries for child call ids are not created for
CS calls in case of conference SRVCC. This causes manage conference
option to not list any participants. To fix this, make sure that
contact info of conference call child call ids are built as soon
as the conference call appears.

Change-Id: I0097ee5e5f760c764214441665a836707cf2f53b
CRs-Fixed: 767767