InCallUI: Update the DialpadFragment state immediately
When we change the fragment state, it will take some seconds not
immediately, then we will get wrong value which is not the new one.
In this issue, we want to get the value of isDialpadVisible()
to dertermin whether show the manage conferrence button. When we
proceed the funciton named ft.commitAllowingStateLoss(), we also get
the old value of the isDialpadVisible().

Call executePendingTransactions() after commitAllowingStateLoss(),
so that it can work immediately, then we can get the right value
of isDialpadVisible().

CRs-Fixed: 771537

Change-Id: If70b0d6be6faf4b2a3669dbf959c74e51e71f080