InCallUI: Hide floating action end call button in incoming call screen
When there is an incoming call, notify all IncomingCallListener
for the incoming call, but due to the delay 500ms to refresh UI,
CallCardPresenter has not register into IncomingCallListener yet,
cause the CallCardPresenter is not notified for incoming call,
therefore the end call button is not hidden.

Set end call button's visibility gone and set enabled false at first.

Change-Id: I3b4d69d2d882decd87bcc1c395cc3ea2f969d482
CRs-Fixed: 885158