IMS-VT: Fix race-condition causing preview freeze.
The VideoCall object is made of two binders -
one for sending requests to telecomm and another one for
receiving notifications from telecom. The initialization of
the latter is delayed since a message is posted on a handler
and initialization is done in the handler. getVideoCall()
of InCallUi's Call class deligates the call to telecom's
getVideoCall() function. Since UI events are asynchronous
request to open the camera is sent, however, if the response
is received before the initialization is complete in the handler
the response will be dropped and camera initialization will fail.

Modify the getVideoCall() function of the InCallUI's call to
return valid VideoCall object only when the object is fully
constructed.

CRs-Fixed: 976997
Change-Id: Id864892bf8452161f2c6f526edd6e4ecc39bf5cd